♻️ Moving window utils into their own module
